Optimizing for bi-directional references in tree
For each node to be able to print it’s path, it must hold references up the tree to it’s parents. To manage risk, logic to add a child is moved inside the node and the parent is kept private. The children are moved to private set and public read. Read available for search purposes.